Extract the test for dsatur
